isPublic icon displayed on offer cards
